re-structure OWIDEnv, wizard.utils -> etl.config
