+++ apps/wizard/app_pages/chart_diff/app.py
@@ -12,8 +12,8 @@ from apps.wizard.app_pages.chart_diff.chart_diff import get_chart_diffs_from_gra
 from apps.wizard.app_pages.chart_diff.chart_diff_show import st_show
 from apps.wizard.app_pages.chart_diff.utils import WARN_MSG, get_engines
 from apps.wizard.utils import Pagination, set_states
-from apps.wizard.utils.env import OWID_ENV
 from etl import config
+from etl.config import OWID_ENV
 
 log = get_logger()
