+++ apps/wizard/app_pages/charts/variable_config.py
@@ -12,7 +12,7 @@ from structlog import get_logger
 from apps.backport.datasync.data_metadata import variable_data_df_from_s3
 from apps.wizard.app_pages.charts.utils import get_variables_from_datasets
 from apps.wizard.utils import set_states
-from apps.wizard.utils.env import OWID_ENV
+from etl.config import OWID_ENV
 from etl.db import get_engine
 from etl.match_variables import find_mapping_suggestions, preliminary_mapping
