+++ apps/wizard/app_pages/indicator_upgrade/charts_update.py
@@ -10,7 +10,7 @@ from structlog import get_logger
 import etl.grapher_model as gm
 from apps.chart_sync.admin_api import AdminAPI
 from apps.wizard.utils import set_states, st_page_link
-from apps.wizard.utils.env import OWID_ENV
+from etl.config import OWID_ENV
 from etl.db import get_engine
 from etl.indicator_upgrade.indicator_update import find_charts_from_variable_ids, update_chart_config
