+++ apps/wizard/app_pages/indicator_upgrade/indicator_mapping.py
@@ -14,7 +14,7 @@ from apps.backport.datasync.data_metadata import variable_data_df_from_s3
 from apps.wizard.app_pages.indicator_upgrade.explore_mode import st_explore_indicator_dialog
 from apps.wizard.app_pages.indicator_upgrade.utils import get_indicators_from_datasets
 from apps.wizard.utils import set_states
-from apps.wizard.utils.env import OWID_ENV
+from etl.config import OWID_ENV
 from etl.db import get_engine
 from etl.match_variables import find_mapping_suggestions, preliminary_mapping
