+++ apps/wizard/cli.py
@@ -13,8 +13,9 @@ import streamlit.web.cli as stcli
 from rich_click.rich_command import RichCommand
 
 from apps.utils.style import set_rich_click_style
-from apps.wizard.config import WIZARD_PHASES, WIZARD_PORT
+from apps.wizard.config import WIZARD_PHASES
 from apps.wizard.utils import CURRENT_DIR
+from etl.config import WIZARD_PORT
 
 # Disable streamlit cache data API logging
 # ref: @kajarenc from https://github.com/streamlit/streamlit/issues/6620#issuecomment-1564735996
