+++ apps/wizard/utils/__init__.py
@@ -30,9 +30,8 @@ from typing_extensions import Self
 
 from apps.wizard.config import PAGES_BY_ALIAS
 from apps.wizard.utils.defaults import load_wizard_defaults, update_wizard_defaults_from_form
-from apps.wizard.utils.env import OWID_ENV, OWIDEnv
 from apps.wizard.utils.step_form import StepForm
-from etl import config
+from etl.config import OWID_ENV, OWIDEnv, enable_bugsnag
 from etl.db import get_connection, read_sql
 from etl.files import ruamel_dump, ruamel_load
 from etl.metadata_export import main as metadata_export
@@ -406,7 +405,7 @@ def _check_env() -> bool:
     """Check if environment variables are set correctly."""
     ok = True
     for env_name in ("GRAPHER_USER_ID", "DB_USER", "DB_NAME", "DB_HOST"):
-        if getattr(config, env_name) is None:
+        if getattr(OWID_ENV.conf, env_name) is None:
             ok = False
             st.warning(f"Environment variable `{env_name}` not found, do you have it in your `.env` file?")
 
@@ -438,10 +437,10 @@ def _show_environment():
     **Environment variables**:
 
     ```
-    GRAPHER_USER_ID: {config.GRAPHER_USER_ID}
-    DB_USER: {config.DB_USER}
-    DB_NAME: {config.DB_NAME}
-    DB_HOST: {config.DB_HOST}
+    GRAPHER_USER_ID: {OWID_ENV.conf.GRAPHER_USER_ID}
+    DB_USER: {OWID_ENV.conf.DB_USER}
+    DB_NAME: {OWID_ENV.conf.DB_NAME}
+    DB_HOST: {OWID_ENV.conf.DB_HOST}
     ```
     """
     )
@@ -600,7 +599,7 @@ def enable_bugsnag_for_streamlit():
     """Enable bugsnag for streamlit. Uses this workaround
     https://github.com/streamlit/streamlit/issues/3426#issuecomment-1848429254
     """
-    config.enable_bugsnag()
+    enable_bugsnag()
     # error_util = sys.modules["streamlit.error_util"]
     error_util = sys.modules["streamlit.runtime.scriptrunner.script_runner"]
     original_handler = error_util.handle_uncaught_app_exception
